backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
python-dotenv
pydantic
pytest
httpx
